Runs revan, but provides more info in a ROOT file
This script is written by Riccardo Rando and is basically a version of revan that allows you to save also information regarding the original simulated events. It is equivalent of running revan, but it creates an additional output in root format. The root file (root_output) will contain several information that are encoded both in the revan and in the cosima step.

How to: 
- open the zip
- load the megalib software
- in the folder where you de-compress the MyAnalysis.zip compile the files, just type “make”

 you should be ready to run:

./myAnalysis -g geofile -s simfile -c revan_cfg_file  -t tra_output -r root_output